+++ src/main/kotlin/cn/yiiguxing/plugin/translate/TranslationPresenter.kt
@@ -23,9 +23,9 @@ class TranslationPresenter(private val view: View, private val recordHistory: Bo
 
     override val primaryLanguage: Lang get() = translator.primaryLanguage
 
-    override val supportedLanguages: SupportedLanguages
+    override val supportedLanguages: SupportedLanguagesData
         get() = with(translator) {
-            SupportedLanguages(supportedSourceLanguages, supportedTargetLanguages)
+            SupportedLanguagesData(supportedSourceLanguages, supportedTargetLanguages)
         }
 
     override fun isSupportedSourceLanguage(sourceLanguage: Lang): Boolean {
@@ -44,6 +44,7 @@ class TranslationPresenter(private val view: View, private val recordHistory: Bo
         return when (settings.targetLanguageSelection) {
             DEFAULT -> Lang.AUTO.takeIf { isSupportedTargetLanguage(it) }
                 ?: if (text.isEmpty() || text.any(NON_LATIN_CONDITION)) Lang.ENGLISH else primaryLanguage
+
             PRIMARY_LANGUAGE -> primaryLanguage
             LAST -> states.lastLanguages.target.takeIf { isSupportedTargetLanguage(it) } ?: primaryLanguage
         }