+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ## [Unreleased]
 
+## [3.6.5] (2024/09/17)
+
+- DeepL Translator now supports languages such as Chinese Traditional.
+- User interaction experience optimization and improvement.
+- Bug fixes.
+- DeepL 翻译现已支持繁体中文等语言
+- 用户交互体验优化与改进
+- Bug 修复
+
 ## [3.6.4] (2024/07/23)
 
 - The GPT-4o mini is now available in the OpenAI Translator engine.
@@ -505,7 +514,8 @@
 - 支持单词拆分。翻译变量名或方法名时更方便
 - Bug修复
 
-[Unreleased]: https://github.com/YiiGuxing/TranslationPlugin/compare/v3.6.4...HEAD
+[Unreleased]: https://github.com/YiiGuxing/TranslationPlugin/compare/v3.6.5...HEAD
+[3.6.5]: https://github.com/YiiGuxing/TranslationPlugin/compare/v3.6.4...v3.6.5
 [3.6.4]: https://github.com/YiiGuxing/TranslationPlugin/compare/v3.6.3...v3.6.4
 [3.6.3]: https://github.com/YiiGuxing/TranslationPlugin/compare/v3.6.2...v3.6.3
 [3.6.2]: https://github.com/YiiGuxing/TranslationPlugin/compare/v3.6.1...v3.6.2